add options of job control builtens

+++ src/job_control/built.c
@@ -13,6 +13,21 @@
 #include "shell.h"
 #include "read_line.h"
 
+void	ft_run_job(t_job *job)
+{
+	t_process	*process;
+	t_list		*proc;
+
+	proc = job->proc;
+	job->status = RUN;
+	while (proc)
+	{
+		process = proc->content;
+		process->status = RUN;
+		proc = proc->next;
+	}
+}
+
 void	ft_update_p_fg(int index)
 {
 	t_list	*tmp;
@@ -35,26 +50,74 @@ void	ft_update_p_fg(int index)
 	}
 }
 
-void	ft_jobs_built(void)
+int		ft_get_jobs_arg(char **args, int *index)
+{
+	int i;
+	int j;
+	int ret;
+
+	if (!args)
+		return (-1);
+	j = 0;
+	ret = -1;
+	while (args[j])
+	{
+		i = 0;
+		if (args[j][0] == '-')
+		{
+			while (args[j][i])
+				i++;
+			if (args[j][i - 1] == 'p')
+				ret = 1;
+			else if (args[j][i - 1] == 'l')
+				ret = 0;
+		}
+		else
+			*index = ft_atoi(args[j]);
+		j++;
+	}
+	return (ret);
+}
+
+void	ft_jobs_built(char **args)
 {
 	t_list	*tmp;
 	t_job	*job;
+	int		index;
+	int		ar;
 
 	tmp = jobs;
+	index = 0;
+	ar = ft_get_jobs_arg(args, &index);
 	while (tmp)
 	{
 		job = tmp->content;
-		ft_putchar('[');
-		ft_putnbr(job->index);
-		ft_putchar(']');
-		(job->p != 0) ? ft_putchar(job->p) : ft_putchar(' ');
-		ft_putstr("  ");
-		ft_printstatus(job->status);
-		ft_putstr("\t\t\t");
-		ft_putstr(job->cmd);
-		ft_putchar(' ');
-		(job->background == 1) ? ft_putchar('&') : 0;
-		ft_putchar('\n');
+		if (index != 0 && job->index != index)
+		{
+			tmp = tmp->next;
+			continue ;
+		}
+		if (ar != 1)
+		{
+			ft_putchar('[');
+			ft_putnbr(job->index);
+			ft_putchar(']');
+			(job->p != 0) ? ft_putchar(job->p) : ft_putchar(' ');
+			ft_putstr("  ");
+			(ar == 0) ? ft_putnbr(job->pgid) : 0;
+			(ar == 0) ? ft_putchar(' ') : 0;
+			ft_printstatus(job->status);
+			ft_putstr("\t\t\t");
+			ft_putstr(job->cmd);
+			ft_putchar(' ');
+			(job->background == 1) ? ft_putchar('&') : 0;
+			ft_putchar('\n');
+		}
+		else if (ar == 1)
+		{
+			ft_putnbr(job->pgid);
+			ft_putchar('\n');
+		}
 		tmp = tmp->next;
 	}
 }
@@ -81,15 +144,15 @@ void	ft_continue(char *arg)
 			(!tmp) ? ft_putendl_fd("42sh: bg: current: no such job", 2) : 0;
 			continue ;
 		}
-		if (job->status == STOPED)
+		if (job->status == STOPED && job->p == '+')
 		{
 			ft_print_backcmd(job);
 			job->background = 1;
-			job->status = RUN;
+			ft_run_job(job);
 			killpg(job->pgid, SIGCONT);
 			break ;
 		}
-		else if (job->status == RUN)
+		else if (job->status == RUN && job->p == '+')
 		{
 			ft_putstr_fd("bash: bg: job ", 2);
 			ft_putnbr(job->index);
@@ -106,7 +169,7 @@ void	ft_foreg_wait(t_job *job, t_list **tmp, t_list **pr)
 		ft_putendl_fd("Controling terminal ERROR", 2);
 	signal(SIGCHLD, SIG_DFL);
 	killpg(job->pgid, SIGCONT);
-	job->status = RUN;
+	ft_run_job(job);
 	job->background = -1;
 	g_sign = 1;
 	ft_wait(job);
@@ -142,6 +205,7 @@ void	ft_foreground(char *arg)
 		job = tmp->content;
 		if (arg && index != job->index)
 		{
+			pr = tmp;
 			tmp = tmp->next;
 			(!tmp) ? ft_putendl_fd("42sh: fg: current: no such job", 2) : 0;
 			continue ;
